test(cli): split sandbox mutation suites
